requirements.txt
pandas
boto3
pyarrow
apache-airflow-providers-postgres